+++ include/NeoN/finiteVolume/cellCentred/fields/volumeField.hpp
@@ -147,80 +147,6 @@ public:
         return boundaryConditions_;
     }
 
-    template<typename T = ValueType>
-        requires std::is_same_v<T, Vec3>
-    [[nodiscard]] VolumeField<scalar> component(const size_t componentIndex) const
-    {
-        return componentField(componentIndex, "component" + std::to_string(componentIndex));
-    }
-
-    template<typename T = ValueType>
-        requires std::is_same_v<T, Vec3>
-    [[nodiscard]] VolumeField<scalar> x() const
-    {
-        return componentField(0, "x");
-    }
-
-    template<typename T = ValueType>
-        requires std::is_same_v<T, Vec3>
-    [[nodiscard]] VolumeField<scalar> y() const
-    {
-        return componentField(1, "y");
-    }
-
-    template<typename T = ValueType>
-        requires std::is_same_v<T, Vec3>
-    [[nodiscard]] VolumeField<scalar> z() const
-    {
-        return componentField(2, "z");
-    }
-
-    template<typename T = ValueType>
-        requires std::is_same_v<T, Vec3>
-    [[nodiscard]] VolumeField<scalar>
-    componentField(const size_t componentIndex, const std::string& suffix) const
-    {
-        if (componentIndex >= 3)
-        {
-            NF_ERROR_EXIT("VolumeField<Vec3>::component index out of range.");
-        }
-
-        VolumeField<scalar> result(
-            this->exec(),
-            this->name + "." + suffix,
-            this->mesh(),
-            createCalculatedBCs<VolumeBoundary<scalar>>(this->mesh())
-        );
-
-        auto sourceInternal = this->internalVector().view();
-        parallelFor(
-            result.internalVector(),
-            NEON_LAMBDA(const localIdx i) { return sourceInternal[i][componentIndex]; }
-        );
-
-        auto sourceValue = this->boundaryData().value().view();
-        parallelFor(
-            result.boundaryData().value(),
-            NEON_LAMBDA(const localIdx i) { return sourceValue[i][componentIndex]; }
-        );
-
-        auto sourceRefValue = this->boundaryData().refValue().view();
-        parallelFor(
-            result.boundaryData().refValue(),
-            NEON_LAMBDA(const localIdx i) { return sourceRefValue[i][componentIndex]; }
-        );
-
-        auto sourceRefGrad = this->boundaryData().refGrad().view();
-        parallelFor(
-            result.boundaryData().refGrad(),
-            NEON_LAMBDA(const localIdx i) { return sourceRefGrad[i][componentIndex]; }
-        );
-
-        result.boundaryData().valueFraction() = this->boundaryData().valueFraction();
-
-        return result;
-    }
-
 private:
 
     std::vector<VolumeBoundary<ValueType>> boundaryConditions_; // The vector of boundary conditions
