+++ src/CMakeLists.txt
@@ -57,9 +57,7 @@ target_sources(
           "finiteVolume/cellCentred/auxiliary/coNum.cpp"
           "timeIntegration/timeIntegration.cpp"
           "timeIntegration/rungeKutta.cpp"
-          "turbulenceModels/ReynoldsStress.cpp"
-          "turbulenceModels/DES/SpalartAllmarasBase.cpp"
-          "turbulenceModels/DES/SpalartAllmarasDDES.cpp")
+          "turbulenceModels/SpalartAllmarasDDES.cpp")
 
 if(NeoN_ENABLE_MPI_SUPPORT)
   target_sources(NeoN PRIVATE "core/mpi/halfDuplexCommBuffer.cpp"
